backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
pydantic>=2.7.0
PyGithub>=2.3.0
google-generativeai>=0.7.0
httpx>=0.27.0
